fix(frontend): Registro validation

+++ frontend/app/src/main/java/com/example/myapplication/PostRegistro.kt
@@ -110,7 +110,7 @@ fun PostRegistros() {
                     relSelection,
                     isShippingCompany,
                     rutPersona.text,
-                    numeroDept.text.toInt(),
+                    numeroDept.text,
                     popUpStringContent
                 )
             },
